+++ packages/at_client_flutter/example/lib/apkam_example.dart
@@ -4,6 +4,7 @@ import 'package:at_auth/at_auth.dart';
 import 'package:at_chops/at_chops.dart';
 import 'package:at_client_flutter/at_client_flutter.dart';
 import 'package:at_lookup/at_lookup.dart';
+import 'package:at_lookup/at_lookup_io.dart';
 import 'package:at_utils/at_logger.dart';
 import 'package:flutter/material.dart';
 
@@ -135,6 +136,13 @@ class _ApkamExamplePageState extends State<ApkamExamplePage> {
         currentAtSign,
         atClient.getPreferences()!.rootDomain,
         atClient.getPreferences()!.rootPort,
+        secondaryAddressFinder: CacheableSecondaryAddressFinder(
+          atClient.getPreferences()!.rootDomain,
+          atClient.getPreferences()!.rootPort,
+        ),
+        transportFactory: SecureSocketTransportFactory(
+          secureSocketConfig: SecureSocketConfig(),
+        ),
       );
 
       // STEP 1: Generate an OTP using the MANAGER client