+++ packages/at_onboarding_cli/lib/src/cli/auth_cli.dart
@@ -9,6 +9,7 @@ import 'package:at_chops/at_chops.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_commons/at_builders.dart';
 import 'package:at_lookup/at_lookup.dart';
+import 'package:at_lookup/at_lookup_io.dart';
 import 'package:at_onboarding_cli/at_onboarding_cli.dart';
 import 'package:at_onboarding_cli/src/util/home_directory_util.dart';
 import 'package:at_utils/at_progress.dart';
@@ -405,6 +406,9 @@ Future<int> status(ArgResults ar) async {
       atSign,
       rootDomain.rootDomain,
       rootDomain.rootPort,
+      secondaryAddressFinder: CacheableSecondaryAddressFinder(rootDomain.rootDomain, rootDomain.rootPort),
+      transportFactory: SecureSocketTransportFactory(
+          secureSocketConfig: SecureSocketConfig()),
     );
     try {
       pk = await al.executeCommand('lookup:publickey$atSign\n', auth: false);